+++ src/main/java/com/gitlab/cdagaming/craftpresence/utils/StringUtils.java
@@ -1072,7 +1072,9 @@ public class StringUtils {
                 Field lookupField = classToAccess.getDeclaredField(currentData.getFirst().toString());
                 lookupField.setAccessible(true);
                 lookupField.set(instance, currentData.getSecond());
-                ModUtils.LOG.debugInfo(ModUtils.TRANSLATOR.translate("craftpresence.logger.info.update.dynamic", currentData.toString(), classToAccess.getName()));
+                if (ModUtils.IS_VERBOSE) {
+                    ModUtils.LOG.debugInfo(ModUtils.TRANSLATOR.translate("craftpresence.logger.info.update.dynamic", currentData.toString(), classToAccess.getName()));
+                }
             } catch (Exception ignored) {
             }
         }